# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Moji API",
    "version": "2.0.0",
    "description": "REST API của Moji. Realtime đi qua Socket.IO và **không** được mô tả ở đây —\nhợp đồng event nằm ở `src/socket/events.js`, có contract test đối chiếu với\n`frontend/src/types/socket.ts`.\n\n**Xác thực.** Access token là JWT đặt ở header `Authorization: Bearer <token>`,\nsống 15 phút. Refresh token là chuỗi ngẫu nhiên đặt trong cookie `refreshToken`\n(httpOnly), được xoay vòng ở mỗi lần dùng; dùng lại một token đã xoay sẽ thu hồi\ntoàn bộ phiên của người đó.\n\n**Lỗi.** Mọi lỗi đều có cùng hình dạng `{ code, message, requestId }`. `code` là\nchuỗi ổn định để client rẽ nhánh; `message` là tiếng Việt để hiển thị và có thể\nđổi bất cứ lúc nào."
  },
  "servers": [
    {
      "url": "http://localhost:5001/api",
      "description": "Máy phát triển"
    },
    {
      "url": "https://{host}/api",
      "description": "Triển khai",
      "variables": {
        "host": {
          "default": "moji.example.com"
        }
      }
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/conversations/{conversationId}": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "summary": "Chi tiết một hội thoại",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6a83cfd55df7471cac7f192a"
            },
            "description": "Id cuộc trò chuyện"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "conversation": {
                      "$ref": "#/components/schemas/Conversation"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Xoá nhóm",
        "description": "Chỉ owner. Xoá mềm; mọi thành viên nhận `conversation:removed` qua socket.",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6a83cfd55df7471cac7f192a"
            },
            "description": "Id cuộc trò chuyện"
          }
        ],
        "responses": {
          "200": {
            "description": "Đã xoá",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/conversations/{conversationId}/seen": {
      "patch": {
        "tags": [
          "Conversations"
        ],
        "summary": "Đẩy con trỏ đã đọc",
        "description": "Server tự suy `lastReadAt` từ `createdAt` đã lưu của tin nhắn được chỉ định — client không gửi thời điểm, nếu không lệch đồng hồ sẽ cho phép đánh dấu đã đọc cả tin nhắn trong tương lai. Con trỏ chỉ đi tới (`$max`), nên một gói tin đến muộn từ tab cũ không kéo lùi được. Đây là bản dự phòng của socket event `read:advance`; cả hai dùng chung service.",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6a83cfd55df7471cac7f192a"
            },
            "description": "Id cuộc trò chuyện"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lastReadMessageId": {
                    "type": "string",
                    "description": "Bỏ trống thì đánh dấu tới tin nhắn cuối"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Con trỏ sau khi đẩy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "lastReadAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "myUnreadCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Thiếu access token, hoặc token sai/hết hạn (`TOKEN_EXPIRED`)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Đã đăng nhập nhưng không đủ quyền (`NOT_A_MEMBER`, `INSUFFICIENT_ROLE`)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Không tìm thấy",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "`VALIDATION_ERROR` — body, query hoặc param không hợp lệ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Mã ổn định để client rẽ nhánh, không đổi theo ngôn ngữ hiển thị",
            "example": "NOT_A_MEMBER"
          },
          "message": {
            "type": "string",
            "description": "Câu tiếng Việt để hiển thị; có thể đổi bất cứ lúc nào"
          },
          "requestId": {
            "type": "string",
            "description": "Đối chiếu với log server"
          }
        }
      },
      "ValidationError": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "type": "object",
            "properties": {
              "details": {
                "type": "object",
                "properties": {
                  "fields": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Tên field → lý do, để form gắn lỗi vào đúng ô"
                  }
                }
              }
            }
          }
        ]
      },
...
      "Participant": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Id người dùng"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "member"
            ],
            "nullable": true
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Nguồn sự thật của trạng thái đã đọc. Biên nhận từng tin được SUY RA từ đây (`lastReadAt >= message.createdAt`) chứ không lưu mảng `readBy` trên mỗi tin — mảng đó sẽ ghi O(số tin × số thành viên) mỗi lần mở một hội thoại tồn đọng."
          }
        }
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "direct",
              "group"
            ]
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Participant"
            }
          },
          "group": {
            "nullable": true,
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "nullable": true
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "avatarUrl": {
                "type": "string",
                "nullable": true
              },
              "createdBy": {
                "type": "string",
                "nullable": true
              }
            }
          },
          "lastMessage": {
            "nullable": true,
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "nullable": true
              },
              "content": {
                "type": "string",
                "nullable": true
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "sender": {
                "nullable": true,
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string",
                    "nullable": true
                  },
                  "avatarUrl": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          },
          "lastMessageAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "unreadCounts": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Số chưa đọc theo từng thành viên"
          },
          "unreadCount": {
            "type": "integer",
            "description": "Số chưa đọc của riêng người gọi — client khỏi phải tra map bằng id của chính mình"
          },
          "myRole": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "member"
            ],
            "nullable": true
          },
          "pinned": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}